pyproject.toml
[tool.poetry]
name = "code translater"
version = "0.1.0"
description = ""
authors = ["Hawks"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytelegrambotapi = "^4.21.0"
telebot = "^0.0.5"
env = "^0.1.0"
python-dotenv = "^1.0.1"
google-generativeai = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
